Inject through java config

+++ eagleboard-api/src/main/java/com/mass3d/common/IdentifiableObjectStore.java
@@ -178,7 +178,7 @@ public interface IdentifiableObjectStore<T> extends GenericStore<T>{
    * @param ids a collection of ids.
    * @return a list of objects.
    */
-  List<T> getById(Collection<Integer> ids);
+  List<T> getById(Collection<Long> ids);
 
   /**
    * Retrieves a list of objects referenced by the given collection of uids.
